+++ app/api/clog.py
@@ -2,10 +2,10 @@ import os
 from django.http import HttpResponse
 from django.conf import settings
 from app.models import BrewPiDevice
-from gravity.models import GravitySensor
+from pathlib import Path
 
 
-def get_filepath_to_log(device_type, logfile="", device_id=None):
+def get_filepath_to_log(device_type, logfile="", device_id=None) -> Path:
     # get_filepath_to_log is being broken out so that we can use it in help/other templates to display which log file
     # is being loaded
     if device_type == "brewpi":
@@ -28,7 +28,7 @@ def get_filepath_to_log(device_type, logfile="", device_id=None):
         return None
 
     # Once we've determined the filename from logfile and device_type, let's open it up & read it in
-    logfile_path = os.path.join(settings.BASE_DIR, 'log', log_filename)
+    logfile_path = settings.ROOT_DIR / 'log' / log_filename
     return logfile_path