+++ fermentrack_django/settings.py
@@ -1,27 +1,51 @@
 import os, sys
+import logging
+
 from django.contrib.messages import constants as message_constants  # For the messages override
 import datetime, pytz, configparser
 #from git import Repo
 import git
+from pathlib import Path
+import datetime, pytz, configparser
 
-from .secretsettings import *  # See fermentrack_django/secretsettings.py.example, or run utils/make_secretsettings.sh
+import environ
+
+ROOT_DIR = Path(__file__).resolve(strict=True).parent.parent
+
+env = environ.Env()
+
+READ_DOT_ENV_FILE = env.bool("DJANGO_READ_DOT_ENV_FILE", default=False)
+if READ_DOT_ENV_FILE:
+    # OS environment variables take precedence over variables from .env
+    env.read_env(str(ROOT_DIR / ".env"))
+
+try:
+    from .secretsettings import *  # See fermentrack_django/secretsettings.py.example, or run utils/make_secretsettings.sh
+except:
+    # If we're running under Docker, there is no secretsettings - everything comes from a .env file
+    SECRET_KEY = env("DJANGO_SECRET_KEY")
+
+USE_DOCKER = env.bool("USE_DOCKER", default=False)
+
 
 # Build paths inside the project like this: os.path.join(BASE_DIR, ...)
 BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 
-DEBUG = True
+DEBUG = env.bool("DJANGO_DEBUG", True)
 
-ALLOWED_HOSTS = ['*']  # This is bad practice, but is the best that we're going to get given our deployment strategy
+# This is bad practice, but is the best that we're going to get given our deployment strategy
+ALLOWED_HOSTS = env.list("DJANGO_ALLOWED_HOSTS", default=['*'])
 
 
 # Check if Sentry is enabled (read in the config filepath)
-CONFIG_INI_FILEPATH = os.path.join(BASE_DIR, 'fermentrack_django', 'config.ini')
+CONFIG_INI_FILEPATH = ROOT_DIR / 'fermentrack_django' / 'config.ini'
 config = configparser.ConfigParser()
 config.read(CONFIG_INI_FILEPATH)
 ENABLE_SENTRY = config.getboolean("sentry", "enable_sentry", fallback=True)
 
+
 try:
-    local_repo = git.Repo(path=BASE_DIR)
+    local_repo = git.Repo(path=ROOT_DIR)
     GIT_BRANCH = local_repo.active_branch.name
 except git.exc.InvalidGitRepositoryError:
     ENABLE_SENTRY = False
@@ -60,12 +84,9 @@ if sys.platform == "darwin":
         INSTALLED_APPS += 'mod_wsgi.server', # Used for the macOS setup
 
 
-# if ENABLE_SENTRY:
-#     import raven
-#     INSTALLED_APPS += 'raven.contrib.django.raven_compat',
-
 MIDDLEWARE = [
     'django.middleware.security.SecurityMiddleware',
+    'whitenoise.middleware.WhiteNoiseMiddleware',
     'django.contrib.sessions.middleware.SessionMiddleware',
     'django.middleware.common.CommonMiddleware',
     'django.middleware.csrf.CsrfViewMiddleware',
@@ -102,20 +123,22 @@ WSGI_APPLICATION = 'fermentrack_django.wsgi.application'
 # Database
 # https://docs.djangoproject.com/en/1.10/ref/settings/#databases
 
-# There are two methods of Docker installs -- Ones using docker-compose which then have the database
-# moved to Postgres, and ones that act as a single image, where we still use sqlite. For the single
-# image versions we move the database to a subdirectory which can then be persisted.
-if os.getenv("USE_DOCKER", default="false").lower() == "true":
-    DB_DIR = os.path.join(BASE_DIR, "db")
-else:
-    DB_DIR = BASE_DIR
 
-DATABASES = {
-    'default': {
-        'ENGINE': 'django.db.backends.sqlite3',
-        'NAME': os.path.join(DB_DIR, 'db.sqlite3'),
-    }
-}
+# There are two methods of Docker installs -- Ones using docker-compose which then have the database
+# moved to Postgres, and ones that act as a single image where we still use sqlite. For the single
+# image versions we move the database to a subdirectory which can then be persisted.
+if USE_DOCKER:
+    DB_DIR = ROOT_DIR / "db"
+else:
+    DB_DIR = ROOT_DIR
+
+# For Docker installs, the environment variable DATABASE_URL is set to load Postgres instead of SQLite
+sqlite_file_location = "sqlite:///" + os.path.join(DB_DIR, 'db.sqlite3')
+
+DATABASES = {"default": env.db("DATABASE_URL", default=sqlite_file_location)}
+DATABASES["default"]["ATOMIC_REQUESTS"] = True  # noqa F405
+DATABASES["default"]["CONN_MAX_AGE"] = env.int("CONN_MAX_AGE", default=60)  # noqa F405
+
 
 
 # Password validation
@@ -155,13 +178,13 @@ USE_TZ = True
 # https://docs.djangoproject.com/en/1.10/howto/static-files/
 
 STATIC_URL = '/static/'
-STATIC_ROOT = 'collected_static'
+STATIC_ROOT = ROOT_DIR / 'staticfiles'
 
 MEDIA_URL = '/media/'
-MEDIA_ROOT = 'media'
+MEDIA_ROOT = ROOT_DIR / 'media'
 
 DATA_URL = '/data/'
-DATA_ROOT = 'data'
+DATA_ROOT = ROOT_DIR / 'data'
 
 
 # Constance configuration
@@ -260,10 +283,7 @@ CONSTANCE_SETUP_URL = 'setup_config'    # Used in @site_is_configured decorator
 
 
 # Redis Configuration (primarily for gravity sensor support)
-REDIS_HOSTNAME = "127.0.0.1"
-REDIS_PORT = 6379
-REDIS_PASSWORD = ""  # Not used for most installations. If you need this, yell in a thread & we'll add to Constance
-
+REDIS_URL = env("REDIS_URL", default=f"redis://127.0.0.1:6379/0")
 
 
 # Huey Configuration
@@ -273,9 +293,7 @@ HUEY = {
     'immediate': False,
 
     'connection': {
-        'host': REDIS_HOSTNAME,
-        'port': REDIS_PORT,
-        'password': REDIS_PASSWORD,  # TODO - Check if this actually works
+        'url': REDIS_URL,
         'read_timeout': 1,
     }
 }
@@ -284,10 +302,20 @@ HUEY = {
 if ENABLE_SENTRY:
     import sentry_sdk
     from sentry_sdk.integrations.django import DjangoIntegration
+    from sentry_sdk.integrations.logging import LoggingIntegration
 
+    SENTRY_DSN = env("SENTRY_DSN", default="http://99c0c3b2c3214cec950891d07ac6b4fb@sentry.optictheory.com:9000/6")
+    SENTRY_LOG_LEVEL = env.int("DJANGO_SENTRY_LOG_LEVEL", logging.INFO)
+
+    sentry_logging = LoggingIntegration(
+        level=SENTRY_LOG_LEVEL,  # Capture info and above as breadcrumbs
+        event_level=logging.ERROR,  # Send errors as events
+    )
+    integrations = [sentry_logging, DjangoIntegration()]
     sentry_sdk.init(
-        dsn="http://99c0c3b2c3214cec950891d07ac6b4fb@sentry.optictheory.com:9000/6",
-        integrations=[DjangoIntegration()],
+        dsn=SENTRY_DSN,
+        integrations=integrations,
+        # environment=env("SENTRY_ENVIRONMENT", default="production"),
         traces_sample_rate=0.0,
         send_default_pii=True,
     )