+++ gravity/views.py
@@ -498,7 +498,7 @@ def gravity_manage(request, sensor_id):
             }
         except ObjectDoesNotExist:
             # The sensor is in an inconsistent state. Delete it.
-            messages.error(request, u"The gravity sensor {} had incomplete configuration and was deleted".format(sensor.name))
+            messages.error(request, f"The gravity sensor {sensor.name} had incomplete configuration and was deleted")
             sensor.delete()
             return redirect("siteroot")
 
@@ -523,10 +523,9 @@ def gravity_manage(request, sensor_id):
             }
         except ObjectDoesNotExist:
             # The sensor is in an inconsistent state. Delete it.
-            messages.error(request, u"The gravity sensor {} had incomplete configuration and was deleted".format(sensor.name))
+            messages.error(request, f"The gravity sensor {sensor.name} had incomplete configuration and was deleted")
             sensor.delete()
             return redirect("siteroot")
-
 
         tilt_coefficient_form = forms.TiltCoefficientForm(initial=initial)
         context['tilt_coefficient_form'] = tilt_coefficient_form
@@ -535,6 +534,9 @@ def gravity_manage(request, sensor_id):
         context['tilt_calibration_points'] = calibration_points
         tilt_calibration_form = forms.TiltGravityCalibrationPointForm(initial={'sensor': sensor.tilt_configuration})
         context['tilt_calibration_form'] = tilt_calibration_form
+
+        tilt_extras = sensor.tilt_configuration.load_extras_from_redis()
+        context['tilt_extras'] = tilt_extras
 
         if sensor.tilt_configuration.connection_type == TiltConfiguration.CONNECTION_BRIDGE:
             # For TiltBridges, we want to give the user the info necessary to configure the device to communicate with
@@ -568,7 +570,7 @@ def almost_json_view(request, sensor_id, log_id):
     # gravity_log = GravityLog.objects.get(id=log_id, device_id=sensor_id)
     gravity_log = GravityLog.objects.get(id=log_id)
 
-    filename = os.path.join(settings.BASE_DIR, settings.DATA_ROOT, gravity_log.full_filename("annotation_json"))
+    filename = settings.ROOT_DIR / settings.DATA_ROOT / gravity_log.full_filename("annotation_json")
 
     if os.path.isfile(filename):  # If there are no annotations, return an empty JsonResponse
         f = open(filename, 'r')