+++ utils/processmgr.py
@@ -12,7 +12,7 @@ import logging
 import argparse
 
 from circus.util import DEFAULT_ENDPOINT_DEALER
-from utils.processmgr_class import ProcessManager
+from .processmgr_class import ProcessManager
 from django.core.exceptions import ObjectDoesNotExist
 
 # Load up the Django specific stuff
@@ -24,6 +24,13 @@ from django.core.wsgi import get_wsgi_application
 application = get_wsgi_application()
 import app.models
 import gravity.models
+
+from fermentrack_django.settings import USE_DOCKER
+
+if USE_DOCKER:
+    CIRCUS_ENDPOINT = "tcp://127.0.0.1:7555"
+else:
+    CIRCUS_ENDPOINT = DEFAULT_ENDPOINT_DEALER
 
 
 LOG = logging.getLogger("processmgr")
@@ -39,10 +46,11 @@ fermentrack_install_location = os.path.split(process_mgr_location)[0]
 
 fermentrack_log_path = os.path.join(fermentrack_install_location, "log")
 
-brewpi_script_path = os.path.join(fermentrack_install_location, "brewpi-script/brewpi.py")
+#brewpi_script_path = os.path.join(fermentrack_install_location, "brewpi-script/brewpi.py")
+brewpi_script_path = "brewpi-script.brewpi"
 
 ########## BrewPi Script Configuration
-BREWPI_SCRIPT_CMD_TEMPLATE = "python -u " + brewpi_script_path + ' --dbcfg "%s"'
+BREWPI_SCRIPT_CMD_TEMPLATE = "python -um " + brewpi_script_path + ' --dbcfg "%s"'
 
 
 def BrewPiDevice_query_db(self):
@@ -59,10 +67,10 @@ def BrewPiDevice_query_db(self):
 ########## Tilt Hydrometer Script Configuration
 
 if sys.platform != "darwin":
-    tilt_monitor_script_path = os.path.join(fermentrack_install_location, "gravity/tilt/tilt_monitor_aio.py")
+    tilt_monitor_script_path = "gravity.tilt.tilt_monitor_aio"
 else:
-    tilt_monitor_script_path = os.path.join(fermentrack_install_location, "gravity/tilt/tilt_monitor_macos.py")
-TILT_SCRIPT_CMD_TEMPLATE = "python -u " + tilt_monitor_script_path
+    tilt_monitor_script_path = "gravity.tilt.tilt_monitor_macos"
+TILT_SCRIPT_CMD_TEMPLATE = "python -um " + tilt_monitor_script_path
 
 
 def TiltConfiguration_query_db(self):
@@ -105,7 +113,7 @@ def run():
         prefix="dev-",
         device_type="BrewPi",
         command_tmpl=BREWPI_SCRIPT_CMD_TEMPLATE,
-        circus_endpoint=DEFAULT_ENDPOINT_DEALER,
+        circus_endpoint=CIRCUS_ENDPOINT,
         logfilepath=fermentrack_log_path,
         log=LOG,
         query_db_func=BrewPiDevice_query_db,
@@ -117,7 +125,7 @@ def run():
         prefix="tilt-",
         device_type="Tilt",
         command_tmpl=TILT_SCRIPT_CMD_TEMPLATE,
-        circus_endpoint=DEFAULT_ENDPOINT_DEALER,
+        circus_endpoint=CIRCUS_ENDPOINT,
         logfilepath=fermentrack_log_path,
         log=LOG,
         query_db_func=TiltConfiguration_query_db,